+++ lib/evil_events/plugins.rb
@@ -4,13 +4,19 @@ module EvilEvents
   # @api public
   # @since 0.2.0
   class Plugins
+    require_relative 'plugins/oj_engine'
+    require_relative 'plugins/ox_engine'
+    require_relative 'plugins/mpacker_engine'
+
     # @since 0.3.0
     extend Shared::ExtensionsMixin
 
-    # @since 0.3.0
-    register_extension(:redis_adapter) { require_relative 'plugins/redis_adapter' }
-    # @since 0.3.0
-    register_extension(:sidekiq_adapter) { require_relative 'plugins/sidekiq_adapter' }
+    # @since 0.5.0
+    register_extension(:oj_engine) { OjEngine.load! }
+    # @since 0.5.0
+    register_extension(:ox_engine) { OxEngine.load! }
+    # @since 0.5.0
+    register_extension(:mpacker_engine) { MpackerEngine.load! }
 
     class << self
       # @param plugins [Symbol,Symbol,Symbol,...]