+++ spec/units/evil_events/core/events/serializers_spec.rb
@@ -15,35 +15,31 @@ describe EvilEvents::Core::Events::Serializers, :stub_event_system do
     it 'has following registered serializers' do
       expect(serializers_container.resolve(:json)).to    be_a(described_class::JSON)
       expect(serializers_container.resolve(:hash)).to    be_a(described_class::Hash)
-      expect(serializers_container.resolve(:xml)).to     be_a(described_class::XML)
-      expect(serializers_container.resolve(:msgpack)).to be_a(described_class::MessagePack)
     end
 
-    specify 'serializers should not be memoized' do
-      expect(serializers_container.resolve(:json)).to    eq(serializers_container.resolve(:json))
-      expect(serializers_container.resolve(:hash)).to    eq(serializers_container.resolve(:hash))
-      expect(serializers_container.resolve(:xml)).to     eq(serializers_container.resolve(:xml))
-      expect(serializers_container.resolve(:msgpack)).to eq(serializers_container.resolve(:msgpack))
+    specify 'serializers should be memoized' do
+      expect(serializers_container.resolve(:json)).to eq(serializers_container.resolve(:json))
+      expect(serializers_container.resolve(:hash)).to eq(serializers_container.resolve(:hash))
     end
 
     specify 'fails when serialization engine cant be recognized', :stub_event_system do
-      system_config.serializers.json.engine = gen_symb
+      system_config.configure { |c| c.serializers.json.engine = gen_symb }
 
       expect { serializers_container.resolve(:json) }.to raise_error(
         EvilEvents::UnrecognizedSerializationEngineError
       )
 
-      system_config.serializers.hashing.engine = gen_symb
+      system_config.configure { |c| c.serializers.hashing.engine = gen_symb }
       expect { serializers_container.resolve(:hash) }.to raise_error(
         EvilEvents::UnrecognizedSerializationEngineError
       )
 
-      system_config.serializers.msgpack.engine = gen_symb
+      system_config.configure { |c| c.serializers.msgpack.engine = gen_symb }
       expect { serializers_container.resolve(:msgpack) }.to raise_error(
         EvilEvents::UnrecognizedSerializationEngineError
       )
 
-      system_config.serializers.xml.engine = gen_symb
+      system_config.configure { |c| c.serializers.xml.engine = gen_symb }
       expect { serializers_container.resolve(:xml) }.to raise_error(
         EvilEvents::UnrecognizedSerializationEngineError
       )